% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{ageFish}
\alias{ageFish}
\title{Age fish from long length frequencies}
\usage{
ageFish(object, longLF, lfname = "length_freq")
}
\arguments{
\item{object}{the fit barrierALK object}

\item{longLF}{the long length frquencies with covariates}

\item{lfname}{name of the length frequency column}
}
\description{
Age fish into a long set of age frequencies instead
}
