% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{anova.barrierALK}
\alias{anova.barrierALK}
\title{LRT testing for barrierALK objects}
\usage{
\method{anova}{barrierALK}(object, ..., dispersion = NULL)
}
\arguments{
\item{object}{barrierALK object to test}

\item{...}{other barrierALK objects to test}
}
\description{
LRT testing for barrierALK objects
}
