% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{barrierALK}
\alias{barrierALK}
\title{Barrier ALK, fit a barrier ALK model to age-growth data to generate ALKs}
\usage{
barrierALK(
  formula,
  minAge,
  maxAge,
  lengthVar = "length",
  equalSlopes = FALSE,
  data = NULL,
  silent = TRUE,
  runSA = TRUE,
  control = NULL,
  ...,
  optMethod = "nlminb",
  cores = parallel::detectCores(),
  lambda = 0,
  penalty.matrix = NULL,
  penalize.hyper = FALSE,
  starting.param = NULL,
  scale = FALSE,
  edf.calc = TRUE
)
}
\arguments{
\item{formula}{formula for the barrier alk fit}

\item{lengthVar}{the name of the length variable in the data}

\item{equalSlopes}{TRUE or FALSE for equal slopes assumption, or formla specifiying which are}

\item{data}{optionally data frame containing data}

\item{silent}{whether to run the trace from TMB}

\item{runSA}{Apply Symbolic Analysis from TMB package in effort to up model}

\item{control}{control options for optimr optimizer}

\item{optMethod}{the optimizer within optimr to use}

\item{lambda}{the value or vector for penalization factors, default is 0, no penalization}

\item{penalty.matrix}{manually set the penalty matrix to use}

\item{penalize.hyper}{penalize the hyper parameters for the gf approximations?}

\item{starting.param}{named list to replace default starting parameter values with}
}
\description{
Fit smooth age-length keys with the possibility of incorporating spatial error terms represented by
approximations of Guassian Fields using SPDEs
}
