% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barrierMesh}
\alias{barrierMesh}
\title{Barrier mesh simplifed for 3Ps}
\format{
An object of class \code{list} of length 3.
}
\usage{
barrierMesh
}
\description{
Barrier mesh simplifed for 3Ps
}
\keyword{datasets}
