% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{createALKs}
\alias{createALKs}
\title{Function to create a smooth ALK from a barrierALK object}
\usage{
createALKs(object, lengthBins, setData)
}
\arguments{
\item{object}{The barrierALK fit object}

\item{lengthBins}{vector containing the length bins you want to generate an age-length key for}

\item{setData}{data you want to apply an ALK to (must contain covariates with same name as in in barrierALK fit)}
}
\description{
Function to create a smooth ALK from a barrierALK object
}
