% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierPlot.R
\name{getALKraw}
\alias{getALKraw}
\title{Generate an emperical ALK from the observed proportions in each age group}
\usage{
getALKraw(age, length, minAge, maxAge, lengthBins, proportion = FALSE)
}
\arguments{
\item{age}{a vector containing the ages}

\item{length}{a vector containing the lengths}

\item{minAge}{pool ages below this together}

\item{maxAge}{pool ages above this together}

\item{lengthBins}{the desired vector of length bins for the ALK}

\item{proportion}{if true gets the ALK in proportion form}
}
\description{
Note the shitty fix for DFO data.
}
