% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{getNatAge}
\alias{getNatAge}
\title{Get numbers at age for a barrierALK object}
\usage{
getNatAge(object, lengthBins, setData, lengthFreq)
}
\arguments{
\item{object}{The fit barrierALK object}

\item{lengthBins}{vector containing the length bins you want to generate numbers at age for}

\item{setData}{data you want to get numbers at age for}

\item{lengthFreq}{length frequencies in same order as set data}
}
\description{
Get numbers at age for a barrierALK object
}
