% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetup.R
\name{gf}
\alias{gf}
\title{Function to handle the gaussian field construction}
\usage{
gf(
  loc_x,
  loc_y,
  model = "barrier",
  mesh = NULL,
  barrier.triangles = NULL,
  rangeKey = NULL,
  sdUkey = NULL,
  ageKey = NULL,
  year = NULL
)
}
\arguments{
\item{loc_x}{The x coordinates of the spatial data}

\item{loc_y}{The y coordinates of the spatial data}

\item{model}{specify whether the model is basic spde or barrier}

\item{mesh}{The mesh used to build the precision matrix}

\item{barrier.triangles}{the set of triangles in the barrier of the mesh, only for barrier model}

\item{rangeKey}{optionally specify which age groups to share ranges in the precision matrices, default is all ages unique}

\item{sdUkey}{optionally specify which age groups to share sigma_u's in the precision matrices, default is all ages unique}

\item{ageKey}{for spatAR, specify age groups share same settings}

\item{year}{factor variable indicating the year to use}
}
\description{
This function specifies the gaussian field used within a spatialALK formula
}
