% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetup.R
\name{interpret.salk}
\alias{interpret.salk}
\title{Interpret a spatial ALK model formula}
\usage{
interpret.salk(sf, equalSlopes = FALSE, data = NULL, minAge, maxAge,
  pgf = NULL)
}
\arguments{
\item{sf}{a spatial ALK model formula}

\item{equalSlopes}{formula or boolean indicating whether to have equal}

\item{data}{the data}

\item{minAge}{the minimum age group to fit}

\item{maxAge}{the maximum age group to fit}

\item{pgf}{the predicted gaussian field workaround}
}
\description{
Reads in a spatial ALK formula and sets up the data for loading into the TMB model.
}
