% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{logLik.barrierALK}
\alias{logLik.barrierALK}
\title{Get the log likelihood of a barrierALK fit}
\usage{
\method{logLik}{barrierALK}(object, ...)
}
\arguments{
\item{object}{a fit barrierALK object}
}
\description{
Get the log likelihood of a barrierALK fit
}
