% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{longALKs}
\alias{longALKs}
\title{Create a long version of the ALKs for merging into length frequeinces, why did I make this?}
\usage{
longALKs(object, lengthBins, setData)
}
\arguments{
\item{object}{The barrierALK fit object}

\item{lengthBins}{vector containing the length bins you want to generate an age-length key for}

\item{setData}{data you want to apply an ALK to}
}
\description{
Create a long version of the ALKs for merging into length frequeinces, why did I make this?
}
