% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierPlot.R
\name{plot.predALK}
\alias{plot.predALK}
\title{Plot a predicted ALK}
\usage{
plot.predALK(x, lwd = 2, type = "l", ylab = "Proportion",
  xlab = "Length (CM)", col = NULL, lty = seq_len(length(colors)),
  ..., rawALK = NULL, plotFreq = FALSE, autoLen = TRUE)
}
\arguments{
\item{x}{the predALK object to plot}

\item{...}{the options to pass onto matplot}

\item{rawALK}{the rawALK proportions
object to plot with the predicted one}
}
\description{
Plot a predicted ALK
}
