% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierPlot.R
\name{plotSpField}
\alias{plotSpField}
\title{Plot a spatial field}
\usage{
plotSpField(field, mesh, xlim, ylim, poly.barrier = NULL, dims = c(300, 300))
}
\arguments{
\item{field}{field to plot}

\item{mesh}{mesh to project on to}

\item{xlim}{x limits of plotting area}

\item{ylim}{y limits of plotting area}

\item{poly.barrier}{optional barrier.polygon to}

\item{dims}{dimension of the projection}
}
\description{
This function returns a ggplot2 plot of a spatial field
}
