% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierPlot.R
\name{PointCorr}
\alias{PointCorr}
\title{The correlation of a precision Q at a single location}
\usage{
PointCorr(Q, location, mesh)
}
\arguments{
\item{Q}{the precision matrix Q to plot}

\item{location}{the location to plot}

\item{mesh}{the mesh used}
}
\description{
The spatial correlation of a single point
}
