% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{predict.barrierALK}
\alias{predict.barrierALK}
\title{Function to predict from a barrierALK fit}
\usage{
\method{predict}{barrierALK}(object, newdata = NULL, type = "response", ...)
}
\arguments{
\item{object}{the barrierALK fit object}

\item{newdata}{optionally the newdata to predict on}

\item{type}{the type of prediction}
}
\description{
This will predict from a barrierALK object with three possible types,
the default response will return the unconditional probability of an
observation belonging to a class, link the etas and class the predicted
class of that observation.
}
