% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{print.barrierALK}
\alias{print.barrierALK}
\title{Print a barrierALK object}
\usage{
\method{print}{barrierALK}(object, digits = 3L, ...)
}
\arguments{
\item{object}{the barrierALK object to print}

\item{digits}{the number of digits to round to}
}
\description{
Print a barrierALK object
}
