% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{print.summary.barrierALK}
\alias{print.summary.barrierALK}
\title{Print a summary of a barrierALK object fit}
\usage{
\method{print}{summary.barrierALK}(object, digits = 3L)
}
\arguments{
\item{object}{the summary.barrierALK object to print}
}
\description{
Print a summary of a barrierALK object fit
}
