% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quickPredict}
\alias{quickPredict}
\title{quickPredict}
\usage{
quickPredict(X, beta, ranX, A, cohort, ranXind)
}
\arguments{
\item{X}{the design matrix}

\item{beta}{the vector fixed effect parameters}

\item{ranX}{the predicted random effects}

\item{A}{the A prediction matrix}

\item{cohort}{group cohort of CRL}

\item{ranXind}{convert cohort to ranX cols}
}
\description{
Quickly predict the output of a spatial ALK model
}
