% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{residuals.barrierALK}
\alias{residuals.barrierALK}
\title{Residuals for the barrierALK object}
\usage{
\method{residuals}{barrierALK}(object, type = "deviance", ...)
}
\arguments{
\item{object}{the barrierALK object to get the residuals of}

\item{type}{the type of residual to return, either deviance, pearson or RQR,PSR}

\item{...}{unused

Residuals of the barrierALK object including support for randomized quantile residuals (RQR) and probability scale residuals (PRS).
Note that for the deviance and pearson residuals there is one or more residuals for each observation from each level of the CRL they pass through. The replicated response that the pearson or deviance residuals actually refer to can be retrived from the model fit along with the indices of how observations were replicated from the original data using fit$yrep and fit$repind respectively.

Randomized Quantile Residuals try to get a single residual for each observation that should be intereptable the same as residuals for a linear model.}
}
\description{
Residuals for the barrierALK object
}
