% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurveyCustom.R
\name{run_stratC}
\alias{run_stratC}
\title{Version of SimSurvey run_strat to accept other strat_data function versionss}
\usage{
run_stratC(sim, strat_data = strat_data())
}
\arguments{
\item{sim}{Simulation from sim_survey}

\item{strat_data}{closure for calculating stratifed data}
}
\description{
Version of SimSurvey run_strat to accept other strat_data function versionss
}
