% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurveyCustom.R
\name{sim_distribution_same}
\alias{sim_distribution_same}
\title{Equally distribute fish among grid cells}
\usage{
sim_distribution_same(sim, grid = make_grid())
}
\arguments{
\item{sim}{A sim object with abundance}

\item{grid}{a grid object}
}
\description{
Equally distribute fish among grid grid cells based on the assumption that
there is no difference in spatial distribution among fish or age or year clustering,etc.
}
