% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{simCRL}
\alias{simCRL}
\title{Simulate CRL data with the specified parameters}
\usage{
simCRL(
  fParamMatrix,
  intercept = TRUE,
  size = 1000,
  hyperp1 = NULL,
  hyperp2 = NULL,
  mesh = NULL,
  barrier.triangles = NULL,
  barrier.polygon = NULL,
  sizeL = size,
  seed = NULL,
  distro = rnorm
)
}
\arguments{
\item{fParamMatrix}{the matrix of fixed parameters, rows are classes-1,columns are number of parameters, first column is
intercept if allowed}

\item{intercept}{use an intercept term?}

\item{size}{number of samples to simulate}

\item{hyperp1}{hyperp1 for the spatial simulation}

\item{hyperp2}{hyperp2 for the spatial simulation}

\item{mesh}{mesh for the spatial simulation}

\item{barrier.triangles}{the triangles defining the barrier for the barrier model}

\item{barrier.polygon}{the barrier polygon to avoid sampling on the barrier}

\item{sizeL}{number of spatial locations to sim, if not given, then unique location for each size}

\item{seed}{the seed to use}

\item{distro}{the distribution to use for the parameters}
}
\value{
A data frame containing the simulated response, any sampled coordinates from the mesh
}
\description{
Simulate CRL ratio data with the specified parameters with the option to add
spatial data based off of a mesh using simFromMesh.
}
