% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{simFromMesh}
\alias{simFromMesh}
\title{Function to simulate data from a mesh given certain hyper-parameters}
\usage{
simFromMesh(mesh, size = 5000, sp_type = "barrier", hyperp1, hyperp2,
  barrier.triangles = NULL, barrier.polygon = NULL, seed = NULL)
}
\arguments{
\item{mesh}{The mesh to simulate from}

\item{size}{the number of locations to sample}

\item{sp_type}{whether spde or barrier type}

\item{hyperp1}{vector of kappas for spde, vector of ranges for barrier}

\item{hyperp2}{vector of taus for spde, vector of sigma_us for barrier}

\item{barrier.triangles}{the triangles defining the barrier for the barrier model}

\item{barrier.polygon}{the barrier polygon to avoid sampling on the barrier}

\item{seed}{The seed for the random number generation}
}
\description{
This simulates spatial data from a mesh from a specified number of precision matrices. The number of
precision matrices is specified by the length of the hyperparameter vectors, which must be equal.
}
