% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurveyCustom.R
\name{strat_data_bALK}
\alias{strat_data_bALK}
\title{Write one for GAM version?
Crusty
Alternative strat_data for barrierALK objects}
\usage{
strat_data_bALK(
  formula,
  b.opts = list(minAge = 1, maxAge = 13, equalSlopes = FALSE, silent = TRUE, runSA =
    TRUE, control = list(eval.max = 5000, iter.max = 5000)),
  mesh = NULL,
  barrier.triangles = NULL,
  cores = parallel::detectCores()
)
}
\arguments{
\item{formula}{the barrierALK formula}

\item{cores}{number of cores to use, on Windows ignored.}

\item{sim}{Simulation from sim_survey}

\item{length_group}{the length group}

\item{bALKopts}{list of options for barrierALK}
}
\description{
This is an alternative for strat_data utilitizing the results of barrierALK
}
