% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurveyCustom.R
\name{strat_data_bClassify}
\alias{strat_data_bClassify}
\title{Instead of making an ALK just classify fish}
\usage{
strat_data_bClassify(formula, b.opts = list(minAge = 1, maxAge = 13,
  equalSlopes = FALSE, silent = TRUE, runSA = TRUE, control = list(eval.max
  = 5000, iter.max = 5000)), mesh = NULL, barrier.triangles = NULL)
}
\arguments{
\item{formula}{the barrierALK formula}

\item{sim}{Simulation from sim_survey}

\item{length_group}{the length group}

\item{bALKopts}{list of options for barrierALK}
}
\description{
Instead of making an ALK just classify fish
}
