% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierALK.R
\name{summary.barrierALK}
\alias{summary.barrierALK}
\title{Summary of a barrierALK object fit}
\usage{
\method{summary}{barrierALK}(object)
}
\arguments{
\item{object}{the barrierALK object to generate a summary of}
}
\description{
Generate a summary of a barrierALK object. This uses and stores a copy of running
TMB::sdreport() on the fit object. This can be a little slow for for some models on the first run.
Gives a summary only of the fixed effects.
}
